+++ CHANGELOG.md
@@ -37,6 +37,12 @@
 - The interface methods `dot()` and `norm()` of ScalarProduct are now `const`. You will
   have to adjust the method signatures in your own scalar product implementations.
 
+- `MultiTypeBlockVector` now implements the interface method `N()`, which
+  returns the number of vector entries.
+
+- `MultiTypeBlockVector` now implements the interface method `dim()`, which
+  returns the number of scalar vector entries.
+
 - `MultiTypeBlockVector::count()` is now `const`
 
 - `SeqILU` can now be used with SIMD data types.
@@ -54,6 +60,12 @@
 
 - The method `getSolverCategory` of `OwnerOverlapCopyCommunication` is deprecated and
   will be removed after Dune 2.7. Use `category()` instead.
+
+- The method `MultiTypeBlockVector::count()` has been deprecated, because its name
+  is inconsistent with the name mandated by the `dune-istl` vector interface.
+
+- The method `MultiTypeBlockMatrix::size()` has been deprecated, because its name
+  is inconsistent with the name mandated by the `dune-istl` vector interface.
 
 # Release 2.6